Finished controllers, fixed repositories, tables, SP. Added new VMs.

+++ src/BookCatalogue/BookCatalogue.Database/dbo/StoredProcedures/USP_Book_Insert.sql
@@ -7,8 +7,8 @@ AS
 
 BEGIN TRANSACTION
 
-	DECLARE @InsertedBook TABLE (Id INT)
-	DECLARE @InsertedBookId INT
+	DECLARE @InsertedBook TABLE (Id BIGINT)
+	DECLARE @InsertedBookId BIGINT
 
 	INSERT INTO [Book]([Title], [Pages], [PublishedDate])
 		OUTPUT INSERTED.Id INTO @InsertedBook
